Refactor worldgen biomes/ores and add admin features - 0.0.43
Major refactor of world generation and ore/biome systems plus server/admin tweaks.

Key changes:
- Reworked biome/ore architecture: Biome now accepts a Seed; BiomeFactory/OreFactory/OreList removed and replaced with a generic EnumFactory to instantiate enums->classes.
- Introduced PerlinOre and simplified Ore API: Ore exposes BlockTransform and TryGenerateOre; ore presence is computed via OrePresentAt; generation now checks IOreReplaceable on blocks.
- Added IOreReplaceable tag and updated Stone/Sandstone/Ice to implement it.
- Moved and reorganized biome/ores interfaces (IHasOre -> Biomes/Interfaces/Ores, IOreNormal, IBiomeInterface) and adjusted biomes to use the new interfaces and per-biome Seed constructors.
- Generator updated to use EnumFactory, new ore flow, sky color gradients, and Utils.ValueFromGradient; cleaned up various helpers and removed old factory code.
- Seed made public, Perlin-based ore provider introduced, and supporting utilities added (Utils.ValueFromGradient).
- Misc: changed Col32.Lerp parameter type, bumped Version from 0.0.42 to 0.0.43, small API/usings cleanup.
- Server/admin changes: Admin and Ban commands now require Admin privilege; added ServerConfig.ElevateHostToAdmin setting with migration; CmdManager now can elevate host to admin via secret codes and persists the config.

Also includes metadata/file moves and Unity meta updates to reflect the restructuring.

+++ Assets/Assemblies/Core/Vectors/Col32.cs
@@ -72,7 +72,7 @@ namespace Larnix.Core.Vectors
 
         public static bool operator !=(Col32 lhs, Col32 rhs) => !(lhs == rhs);
 
-        public static Col32 Lerp(Col32 start, Col32 end, float t = 0.5f)
+        public static Col32 Lerp(Col32 start, Col32 end, double t = 0.5f)
         {
             t = Math.Clamp(t, 0f, 1f);
             byte lerp(byte a, byte b) => (byte)(a + (b - a) * t);
